requirements.txt
streamlit
google-generativeai
python-dotenv
speechrecognition
pyttsx3
pyaudio
pypdf
python-docx